Update check_running() function in refine
When we are on strict shell system, The "$?" system variable is reset to zero when it is evaluated by the " boolean if" statement. 
It compares "$?" to "7", resets "$?", and then evaluates "$?" to "22".
The "22" case is then not evaluated. 
It causes the "NOT_RUNNING" variable to be ignored when it shouldn't, what leads to the error:
"Something is already running on http://127.0.0.1:3333/"

To avoid this issue, the solution is to attribute its value to a user variable and then, evaluate the variable.